# Conflicts:
#	backend/src/main/resources/config/liquibase/master.xml
#	payroll-accounting/codegen.ts
#	payroll-accounting/graphql/gql/graphql.ts
#	payroll-accounting/styles/globals.css

+++ backend/src/main/groovy/com/backend/gbp/rest/dto/JobItemsDto.groovy
@@ -12,7 +12,7 @@ class JobItemsDto {
     String service
     String serviceCategory
     String descriptions
-    Integer qty
+    BigDecimal qty
     BigDecimal cost
     BigDecimal subTotal
     BigDecimal outputTax
